backend/pyproject.toml
[project]
name = "fintrack"
version = "0.1.0"
description = ""
authors = [
    {name = "me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.1.0,<8.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "cryptography (>=46.0.3,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
